# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DocPatient Nexus API",
    "version": "2.0.0",
...
    "contact": {
      "name": "Miyaad Shah Joy",
      "email": "docpatientnexus@example.com"
    }
  },
  "servers": [
    {
      "url": "https://docpatient-nexus.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/api/v2/patients/payments/checkout-session": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Create a checkout session for an appointment.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "createCheckoutSession",
        "description": "Allows a patient to create a checkout session for an appointment. The patient must be `logged in` to use this route. Log in with a valid `jwt` token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "appointment"
                ],
                "properties": {
                  "appointment": {
                    "type": "string",
                    "example": "64b8e3b9c1b2c4b8e3b9c1b2"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Checkout session created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Checkout session created successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "session": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "example": "cs_test_a15j3Mq8p9KIXYIHUob4TiPEf1vJZZdIvLB2XBubeTMhU3Ig44on4WBhn2"
                            },
                            "object": {
                              "type": "string",
                              "example": "checkout.session"
                            },
                            "amount_subtotal": {
                              "type": "integer",
                              "example": 80000
                            },
                            "amount_total": {
                              "type": "integer",
                              "example": 80000
                            },
                            "currency": {
                              "type": "string",
                              "example": "bdt"
                            },
                            "customer_email": {
                              "type": "string",
                              "example": "farzana.nahar@example.com"
                            },
                            "client_reference_id": {
                              "type": "string",
                              "description": "The appointment ID"
                            },
                            "cancel_url": {
                              "type": "string",
                              "format": "uri",
                              "example": "http://localhost:3000"
                            },
                            "success_url": {
                              "type": "string",
                              "format": "uri",
                              "example": "http://127.0.0.1:3000"
                            },
                            "expires_at": {
                              "type": "integer",
                              "description": "Unix timestamp of session expiration",
                              "example": 1750510918
                            },
                            "livemode": {
                              "type": "boolean",
                              "example": false
                            },
                            "mode": {
                              "type": "string",
                              "example": "payment"
                            },
                            "payment_status": {
                              "type": "string",
                              "example": "unpaid"
                            },
                            "payment_method_types": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              },
                              "example": [
                                "card"
                              ]
                            },
                            "url": {
                              "type": "string",
                              "format": "uri",
                              "description": "Stripe-hosted checkout URL",
                              "example": "https://checkout.stripe.com/c/pay/cs_test_a15j3Mq8p9KIXYIHUob4TiPE..."
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Invalid appointment ID.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "This appointment has already been paid for."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access. Only logged-in patients can access this route. Log in with a valid `jwt` token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "You are not authorized to access this route. Please log in."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden access. Only logged-in patients can access this route.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "You do not have permission to perform this action."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Appointment not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "No appointment is booked with this ID"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error. Something went wrong while processing your request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Something went wrong. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}